move tests to {comp}/test
