+++ plop-templates/Component/Component.test.tsx.hbs
@@ -1,7 +1,7 @@
 import React from "react";
 import "@testing-library/jest-dom";
 import { render } from "@testing-library/react";
-import {{pascalCase name}} from './{{pascalCase name}}';
+import {{pascalCase name}} from '../{{pascalCase name}}';
 
 describe("{{pascalCase name}}", () => {
   it("renders without error", () => {
