+++ src/components/Button/test/Button.test.tsx
@@ -1,10 +1,8 @@
 import React from "react";
 import "@testing-library/jest-dom";
 import { render } from "@testing-library/react";
-import Brand from './Brand';
+import Button from "../Button";
 
-describe("Brand", () => {
-  it("renders without error", () => {
-
-  });
+describe("Button", () => {
+  it("renders without error", () => {});
 });
