+++ src/components/Card/test/Card.test.tsx
@@ -1,10 +1,8 @@
 import React from "react";
 import "@testing-library/jest-dom";
 import { render } from "@testing-library/react";
-import Card from './Card';
+import Card from "../Card";
 
 describe("Card", () => {
-  it("renders without error", () => {
-
-  });
+  it("renders without error", () => {});
 });
