+++ src/components/Header/test/Header.test.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import "@testing-library/jest-dom";
 import { render } from "@testing-library/react";
-import Header from "./Header";
+import Header from "../Header";
 
 describe("Header", () => {
   it("renders brand element", () => {
