+++ src/components/ProjectSearch/test/ProjectSearch.test.tsx
@@ -1,10 +1,8 @@
 import React from "react";
 import "@testing-library/jest-dom";
 import { render } from "@testing-library/react";
-import ProjectSearch from './ProjectSearch';
+import ProjectSearch from "../ProjectSearch";
 
 describe("ProjectSearch", () => {
-  it("renders without error", () => {
-
-  });
+  it("renders without error", () => {});
 });
